Fix failing unit tests: decimal parsing, nullable omission, test races
- DefaultInjector: decimal/decimal? ParseBody used long.Parse instead of
  decimal.Parse, throwing FormatException on non-integer or large decimals.
- ClassSourceProducer: nullable value-type members (bool?, int?, decimal?,
  DateTime?, Guid?) are now omitted entirely on serialize when null,
  matching the existing behavior for reference-type members instead of
  emitting empty tags.
- ComplexFixture/ComplexFixtureV2: both share a process-wide cached
  InfoContainer instance (XmlFactory reuse feature) and were racing when
  xUnit ran them in different parallel collections; grouped them into one
  xUnit collection so they run sequentially.
- CoverageExpansionFixture: corrected two new tests whose expectations
  didn't match the library's actual (and correct) contract - element tags
  use property names (not XSD type names) outside of collections, and the
  length estimator is documented to over-estimate rather than match exactly.

+++ XmlSerDe.Components/Injector/DefaultInjector.cs
@@ -483,12 +483,12 @@ namespace XmlSerDe.Components.Injector
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void ParseBody(roschar body, out decimal result)
         {
-            result = long.Parse(body);
+            result = decimal.Parse(body);
         }
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void ParseBody(roschar body, out decimal? result)
         {
-            result = long.Parse(body);
+            result = decimal.Parse(body);
         }
